normatizado, com expansao das variaveis parcial
