+++ include/tokenizer.h
@@ -6,7 +6,7 @@
 /*   By: aalcara- <aalcara-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/04 15:02:27 by aalcara-          #+#    #+#             */
-/*   Updated: 2021/09/12 10:51:08 by aalcara-         ###   ########.fr       */
+/*   Updated: 2021/09/12 13:43:24 by aalcara-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,10 +60,18 @@ typedef struct s_tokendata
 	char 		buffer[1000];
 }t_tokendata;
 
+t_token	*tokenizer(char *input);
+void	get_token(t_tokendata *tk, t_token **tokens);
+void	lexer(t_token *token);
 void	free_tokens(t_token **tokens);
-t_token	*tokenizer(char *input);//
+//utils
+char	*expand_word(char *key, int len);
+bool	check_is_closed(char *str, char quote);
+char	*ft_strjoinrealloc(char *s1, char *s2, int len);
+bool	special_token(t_token *token, const char *str);
 
-
+//parser
 void	parse_tokens(t_token *tokens);
-
+void	command_parser(t_token *tokens, t_token *pipe, int *last_pipe_in);
+char	**create_command_array(t_token *start, t_token *pipe);
 #endif
