+++ src/lsh_loop.c
@@ -6,7 +6,7 @@
 /*   By: aalcara- <aalcara-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/07 21:00:48 by rdutenke          #+#    #+#             */
-/*   Updated: 2021/09/12 10:46:10 by aalcara-         ###   ########.fr       */
+/*   Updated: 2021/09/12 13:44:54 by aalcara-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,9 +28,9 @@ void	free_array(char **array)
 void	lsh_loop(void)
 {
 	char	*line;
-	// char	**args;
 	int		status;
 	t_token	*tokens;
+	// char	**args;
 
 	status = 1;
 	while (status)
