+++ src/token_free.c
@@ -6,7 +6,7 @@
 /*   By: aalcara- <aalcara-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/12 10:45:12 by aalcara-          #+#    #+#             */
-/*   Updated: 2021/09/12 10:45:31 by aalcara-         ###   ########.fr       */
+/*   Updated: 2021/09/12 11:14:50 by aalcara-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
